Improve movement and debug services
* Fix path smoothing, dll loader, and add hs nodes

- Improve pathfinder so it can take a radius instead of operating on the
bot's radius only.
- Improve steering by setting the max speed to wormhole adjusted speed
when speed is detected over max.
- Add hyperspace specific nodes for moving to base and defending.
- Fix raycaster bug where it wasn't aligning to tiles properly.
- Fix loader so it waits until the main dll is unloaded before copying.

* Implement new steering actuator and debug services

- Add simple line and text rendering for debugging.
- Implement a new steering actuator that puts constraints on target
facing orientation.

* Improve text rendering and out-of-center warping

- Add a flag for centering rendered text.
- Add rendered force vector for debugging.
- Move out-of-center warping to a behavior node instead of being in the
update loop.
- Fix GetEnergy in ContinuumGameProxy so it returns the energy.

* Fix hovering and adjust values

- Fix the seek target when the target distance is higher than the
current distance.
- Tune hovering distances.
- Change aggression so it can decrease when energy is low.
- Change bundle shots to always point at the target exactly.
- Adjust dodge distance to 40 from 25.

+++ Marvin/Player.h
@@ -22,7 +22,7 @@ struct Player {
   uint16_t frequency;
   uint16_t discrete_rotation;
 
-  uint16_t energy;
+  int32_t energy;
   uint8_t status;
 
   Vector2f GetHeading() const {
